+++ Type4Me/ASR/Providers/DeepgramASRConfig.swift
@@ -9,6 +9,8 @@ struct DeepgramASRConfig: ASRProviderConfig, Sendable {
 
     static let supportedModels = [
         "nova-3",
+        "nova-3-general",
+        "nova-3-medical",
         "nova-2",
         "nova-2-general",
         "nova-2-meeting",